drivers: Extract periph_memory_alloc interface
