+++ src/drivers/net/dwc_gmac/Mybuild
@@ -20,4 +20,5 @@ module dwc_gmac {
 	depends embox.driver.net.phy
 	depends embox.net.entry_api
 	depends embox.net.util.show_packet
+	@NoRuntime depends embox.driver.periph_memory_alloc
 }
